Add webq_rs to android

+++ rust/src/lib.rs
@@ -3,7 +3,7 @@ use uniffi;
 uniffi::setup_scaffolding!();
 
 #[uniffi::export]
-pub fn add(left: u64, right: u64) -> u64 {
+pub fn add(left: i32, right: i32) -> i32 {
     left + right
 }
